move deployment related stuff to a dedicated subpackage
