+++ mmdet/models/anchor_heads/anchor_head.py
@@ -253,17 +253,17 @@ class AnchorHead(nn.Module):
             >>> assert len(det_bboxes) == len(det_labels) == cfg.max_per_img
         """
         from torch.onnx import operators, is_in_onnx_export
+        from ...utils.deployment import TracerStub
 
         assert len(cls_scores) == len(bbox_preds)
         num_levels = len(cls_scores)
-
-        # FIXME. Workaround for OpenVINO-friendly export.
-        #        cls_scores[i] is_in_onnx_export()
-        in_export = is_in_onnx_export()
         device = cls_scores[0].device
+        # FIXME. Workaround for OpenVINO-friendly export:
+        #        passing scores tensor itself instead of its spatial size.
+        pass_tensor = is_in_onnx_export() and isinstance(self.anchor_generators[0].grid_anchors, TracerStub)
         mlvl_anchors = [
             self.anchor_generators[i].grid_anchors(
-                cls_scores[i] if in_export else cls_scores[i].size()[-2:],
+                cls_scores[i] if pass_tensor else cls_scores[i].size()[-2:],
                 self.anchor_strides[i],
                 device=device) for i in range(num_levels)
         ]
