+++ mmdet/utils/__init__.py
@@ -1,5 +1,4 @@
 from .flops_counter import get_model_complexity_info
 from .registry import Registry, build_from_cfg
-from .tracer_stubs import TracerStub
 
-__all__ = ['Registry', 'build_from_cfg', 'get_model_complexity_info', 'TracerStub']
+__all__ = ['Registry', 'build_from_cfg', 'get_model_complexity_info']
