+++ mmdet/utils/deployment/symbolic.py
@@ -1,0 +1,87 @@
+import torch
+import torch.onnx.symbolic_helper as sym_help
+from torch.onnx.symbolic_helper import parse_args
+from torch.onnx.symbolic_registry import register_op
+
+
+@parse_args('v', 'v', 'v', 'v', 'none')
+def addcmul_symbolic(g, self, tensor1, tensor2, value=1, out=None):
+    from torch.onnx.symbolic_opset9 import add, mul
+
+    if out is not None:
+        sym_help._unimplemented("addcmul", "Out parameter is not supported for addcmul")
+
+    x = mul(g, tensor1, tensor2)
+    value = sym_help._maybe_get_scalar(value)
+    if sym_help._scalar(value) != 1:
+        value = sym_help._if_scalar_type_as(g, value, x)
+        if not sym_help._is_value(value):
+            value = g.op(
+                "Constant", value_t=torch.tensor(value, dtype=torch.float32))
+        x = mul(g, x, value)
+    return add(g, self, x)
+
+
+def view_as_symbolic(g, self, other):
+    from torch.onnx.symbolic_opset9 import reshape_as
+    return reshape_as(g, self, other)
+
+
+@parse_args('v', 'v', 'i', 'i', 'i', 'none')
+def topk_symbolic(g, self, k, dim, largest, sorted, out=None):
+
+    def reverse(x):
+        from torch.onnx.symbolic_opset9 import reshape, transpose, size
+
+        y = transpose(g, x, 0, dim)
+        shape = g.op("Shape", y)
+        y = reshape(g, y, [0, 1, -1])
+        n = size(g, y, g.op("Constant", value_t=torch.LongTensor([0])))
+        y = g.op("ReverseSequence", y, n, batch_axis_i=1, time_axis_i=0)
+        y = reshape(g, y, shape)
+        y = transpose(g, y, 0, dim)
+        return y
+
+    if out is not None:
+        sym_help._unimplemented("TopK", "Out parameter is not supported for topk")
+    k = sym_help._maybe_get_const(k, 'i')
+    if not sym_help._is_value(k):
+        k = g.op("Constant", value_t=torch.tensor(k, dtype=torch.int64))
+    from torch.onnx.symbolic_opset9 import unsqueeze
+    k = unsqueeze(g, k, 0)
+    top_values, top_indices = g.op("TopK", self, k, axis_i=dim, outputs=2)
+    if not largest:
+        top_values = reverse(top_values)
+        top_indices = reverse(top_indices)
+    return top_values, top_indices
+
+
+@parse_args('v', 'i', 'v', 'v', 'f', 'i')
+def group_norm_symbolic(g, input, num_groups, weight, bias, eps, cudnn_enabled):
+    from torch.onnx.symbolic_opset9 import reshape, mul, add, reshape_as
+
+    channels_num = input.type().sizes()[1]
+
+    if num_groups == channels_num:
+        output = g.op('InstanceNormalization', input, weight, bias, epsilon_f=eps)
+    else:
+        # Reshape from [n, g * cg, h, w] to [1, n * g, cg * h, w].
+        x = reshape(g, input, [0, num_groups, -1, 0])
+        x = reshape(g, x, [1, -1, 0, 0])
+        # Normalize channel-wise.
+        x = g.op('MeanVarianceNormalization', x, axes_i=[2, 3])
+        # Reshape back.
+        x = reshape_as(g, x, input)
+        # Apply affine transform.
+        x = mul(g, x, reshape(g, weight, [1, channels_num, 1, 1]))
+        output = add(g, x, reshape(g, bias, [1, channels_num, 1, 1]))
+
+    return output
+
+
+def register_extra_symbolics(opset=10):
+    assert opset >= 10
+    register_op("addcmul", addcmul_symbolic, "", opset)
+    register_op("view_as", view_as_symbolic, "", opset)
+    register_op("topk", topk_symbolic, "", opset)
+    register_op("group_norm", group_norm_symbolic, "", opset)
